fixed training issues

+++ FExperiment/NN/Connection.fs
@@ -1,8 +1,8 @@
 ﻿module Connection
 
 type public Connection(weight) = 
-    member val value: double = 0.0 with get, set
-    member val weight: double = 0.0 with get, set
+    member val value: float = 0.0 with get, set
+    member val weight: float = 0.0 with get, set
     
-    member this.push(value: double) = 
+    member this.push(value: float) = 
         this.value <- value;
