+++ FExperiment/NN/Layers/InputLayer.fs
@@ -8,7 +8,7 @@ open Neuron
 type InputLayer() = 
     inherit Layer()
 
-    member this.set(value: List<double>) =
+    member this.set(value: List<float>) =
         let self = (this:>ILayer<Neuron>)
         if value.Length = self.neurons.Length then printf("test")
